# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IBKR Market Data API",
    "description": "A web service that connects to Interactive Brokers (IBKR) and provides historical market data through a REST API.",
    "version": "1.0.0",
    "contact": {
      "name": "IBKR Market Data API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api": {
      "get": {
        "summary": "API Documentation",
        "description": "Swagger/OpenAPI documentation interface",
        "responses": {
          "200": {
            "description": "API documentation page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": ["Documentation"]
      }
    },
    "/marketdata": {
      "post": {
        "summary": "Get historical bar data for a symbol",
        "description": "Returns historical bar data (OHLC) for a given stock symbol with customizable parameters",
        "tags": ["Market Data"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDataRequestDto"
              },
              "examples": {
                "simple": {
                  "value": {
                    "symbol": "AAPL",
                    "duration": "2 D",
                    "barSize": "5 mins"
                  }
                },
                "full": {
                  "value": {
                    "symbol": "MSFT",
                    "secType": "STK",
                    "endDateTime": "20230630-16:00:00",
                    "duration": "3 D",
                    "barSize": "15 mins",
                    "whatToShow": "TRADES",
                    "useRTH": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Historical bar data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BarDataDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Symbol is required"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Contract not found"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Failed to connect to IBKR"
                }
              }
            }
          }
        }
      }
    },
    "/historicalticks": {
      "post": {
        "summary": "Get historical tick data for a symbol",
        "description": "Returns historical tick-by-tick data for a given stock symbol with customizable parameters",
        "tags": ["Market Data"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalTicksRequestDto"
              },
              "examples": {
                "simple": {
                  "value": {
                    "symbol": "AAPL"
                  }
                },
                "full": {
                  "value": {
                    "symbol": "AAPL",
                    "secType": "STK",
                    "startDate": "20230101-00:00:00",
                    "endDate": "20230131-23:59:59",
                    "numberOfTicks": 500,
                    "useRTH": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Historical tick data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TickDataDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Symbol is required"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Contract not found"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "error": "Failed to connect to IBKR"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MarketDataRequestDto": {
        "type": "object",
        "required": ["symbol"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Stock symbol (e.g., AAPL, MSFT)",
            "example": "AAPL"
          },
          "secType": {
            "type": "string",
            "description": "Security type",
            "default": "STK",
            "example": "STK"
          },
          "endDateTime": {
            "type": "string",
            "description": "End date and time in format 'yyyymmdd-hh:mm:ss'",
            "example": "20230630-16:00:00"
          },
          "duration": {
            "type": "string",
            "description": "Duration string (e.g., '1 D', '5 D', '1 M')",
            "default": "1 D",
            "example": "2 D"
          },
          "barSize": {
            "type": "string",
            "description": "Bar size setting (e.g., '1 min', '5 mins', '1 hour')",
            "default": "1 min",
            "example": "5 mins"
          },
          "whatToShow": {
            "type": "string",
            "description": "Type of data to retrieve (e.g., 'TRADES', 'MIDPOINT')",
            "default": "TRADES",
            "example": "TRADES"
          },
          "useRTH": {
            "type": "boolean",
            "description": "Use regular trading hours only",
            "default": true,
            "example": true
          }
        }
      },
      "HistoricalTicksRequestDto": {
        "type": "object",
        "required": ["symbol"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Stock symbol (e.g., AAPL, MSFT)",
            "example": "AAPL"
          },
          "secType": {
            "type": "string",
            "description": "Security type",
            "default": "STK",
            "example": "STK"
          },
          "startDate": {
            "type": "string",
            "description": "Start date and time in format 'yyyymmdd-hh:mm:ss', defaults to 1 month ago",
            "example": "20230101-00:00:00"
          },
          "endDate": {
            "type": "string",
            "description": "End date and time in format 'yyyymmdd-hh:mm:ss'",
            "example": "20230131-23:59:59"
          },
          "numberOfTicks": {
            "type": "integer",
            "description": "Maximum number of ticks to retrieve",
            "default": 1000,
            "example": 500
          },
          "useRTH": {
            "type": "boolean",
            "description": "Use regular trading hours only",
            "default": true,
            "example": true
          }
        }
      },
      "BarDataDto": {
        "type": "object",
        "required": ["date", "open", "high", "low", "close", "volume"],
        "properties": {
          "date": {
            "type": "string",
            "description": "Date and time of the bar",
            "example": "2023-06-30T16:00:00.000Z"
          },
          "open": {
            "type": "number",
            "description": "Opening price",
            "example": 185.34
          },
          "high": {
            "type": "number",
            "description": "Highest price during the bar",
            "example": 186.12
          },
          "low": {
            "type": "number",
            "description": "Lowest price during the bar",
            "example": 185.21
          },
          "close": {
            "type": "number",
            "description": "Closing price",
            "example": 185.92
          },
          "volume": {
            "type": "integer",
            "description": "Trading volume",
            "example": 12500
          }
        }
      },
      "TickDataDto": {
        "type": "object",
        "required": ["time", "price", "size"],
        "properties": {
          "time": {
            "type": "string",
            "description": "Time of the tick",
            "example": "2023-01-15T14:30:45.000Z"
          },
          "price": {
            "type": "number",
            "description": "Price of the tick",
            "example": 185.75
          },
          "size": {
            "type": "integer",
            "description": "Size of the tick",
            "example": 100
          }
        }
      },
      "ErrorDto": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Symbol is required"
          }
        }
      }
    }
  }
}